CP-32124: Check if the CRLpath is an empty directory
In stunnel-4.56, CRLpath can be empty directory, as stunnel could ignore
the CRL (certificate revocation list) check if it found an empty
CRLpath.

In stunnel-5.55, this checking logic is moved into OpenSSL, in which an
empty CRLpath will cause checking error.

So in upgrading to stunnel-5.55, CRLpath needs to be checked if it is an
empty directory before specifying it to stunnel.

Signed-off-by: Ming Lu <ming.lu@citrix.com>